모바일개발을 위한 구조 변경 및 일부 리팩터링, Popup forwardRef 제거(대신 redux 사용하도록 리팩터링)
