+++ story-front/components/Footer/Mobile/index.tsx
@@ -1,0 +1,8 @@
+const Footer: React.FC = () => {
+    return (
+        <div>
+        </div>
+    )
+}
+
+export default Footer
