+++ story-front/components/Footer/constant.ts
@@ -1,0 +1,13 @@
+const COMPANY_INTRO = '회사소개'
+const PRIVACY_POLICY = '개인정보처리방침'
+const TEL = '1588 - 0475'
+const WORK_TIME = '월-금 09:00 - 17:00'
+const REST_TIME = '토-일'
+const OWNER = '임원재'
+const REGISTER_NUMBER = '768-15-01230'
+const COMPANY_NAME = '스토리창'
+const FACTORY_ADDR = '경기도 광주시 오포읍 매자리길 53-32'
+const DESC = '스토리 창은 소매업체를 적극 지원하는 전국 블라인드 도매업체입니다.'
+
+export { COMPANY_INTRO, PRIVACY_POLICY, TEL, WORK_TIME, REST_TIME,
+    OWNER, REGISTER_NUMBER, COMPANY_NAME, FACTORY_ADDR, DESC }
