+++ story-front/components/Header/Mobile/index.tsx
@@ -1,0 +1,9 @@
+const Header: React.FC = () => {
+
+	return (
+		<div>
+		</div>
+	)
+}
+
+export default Header
