+++ story-front/components/Header/Pc/index.module.scss
