+++ story-front/components/Popup/Pc/Footer/index.tsx
