+++ story-front/store/action/reducerAction.ts
@@ -23,10 +23,56 @@ export abstract class ReducerActionImpl<T> implements ReducerAction {
     }
 }
 
+
 export interface Data {
     loading: boolean;
     done: boolean;
 }
+
+
+export interface PopupData {
+    visible?: boolean;
+}
+
+export interface ImageData {
+    img_url?: string;
+}
+
+export interface PopupImageData extends PopupData, ImageData {}
+
+export interface CollectPopupData {
+    collect_popup?: PopupData;
+}
+
+export interface ImagePopupData {
+    image_popup?: PopupImageData;
+}
+
+
+export class CollectPopupSetData extends ReducerActionImpl<PopupData> {
+    constructor(data: PopupData){
+        super('SET_COLLECT_POPUP', data);
+    }
+
+    public updateState(state: {}) {
+        return {
+            ...state, collect_popup: { ...this.data}
+        }
+    }
+}
+
+export class ImagePopupSetData extends ReducerActionImpl<PopupImageData> {
+    constructor(data: PopupImageData){
+        super('SET_IMAGE_POPUP', data);
+    }
+
+    public updateState(state: ImagePopupData) {
+        return {
+            ...state, image_popup: {...state?.image_popup, ...this.data }
+        }
+    }
+}
+
 
 type Product = {
     id: number;
@@ -45,6 +91,7 @@ export interface ProductListData extends Data {
     remain?: boolean;
     tag?: string;
 }
+
 
 export class ResponseProductListAction extends ReducerActionImpl<ProductListData> {
     constructor(data: ProductListData){
